GUACAMOLE-1867: Migrate common convenience library to new memory management functions.

+++ src/common/common/string.h
@@ -22,24 +22,36 @@
 
 #include "config.h"
 
+#include <stddef.h>
+
 /**
  * Counts the number of occurrences of a given character in a string.
  *
- * @param string The string to count occurrences within.
- * @param c The character to count occurrences of.
- * @return The number of occurrences.
+ * @param string
+ *     The string to count occurrences within.
+ *
+ * @param c
+ *     The character to count occurrences of.
+ *
+ * @return
+ *     The number of occurrences.
  */
-int guac_count_occurrences(const char* string, char c);
+size_t guac_count_occurrences(const char* string, char c);
 
 /**
  * Splits a string into a newly-allocated array of strings. The array itself
- * and each string within the array will eventually need to be freed. The array
- * is NULL-terminated.
+ * and each string within the array will eventually need to be freed through
+ * calls to guac_mem_free(). The array is NULL-terminated.
  *
- * @param string The string to split.
- * @param delim The character which separates individual substrings within the
- *              given string.
- * @return A newly-allocated, NULL-terminated array of strings.
+ * @param string
+ *     The string to split.
+ *
+ * @param delim
+ *     The character which separates individual substrings within the
+ *     given string.
+ *
+ * @return
+ *     A newly-allocated, NULL-terminated array of strings.
  */
 char** guac_split(const char* string, char delim);
